+++ src/Data/SExpresso/Parse/Location.hs
@@ -9,11 +9,15 @@
 -- The module "Data.SExpresso.Parse" re-exports the functions and
 -- datatypes of this module.
 
+{-# LANGUAGE DeriveFunctor #-}
+
 module Data.SExpresso.Parse.Location
   (
     Location(..),
     Located(..),
-    located
+    located,
+    startPosPretty,
+    endPosPretty
    )
   where
 
@@ -25,9 +29,17 @@ import Text.Megaparsec
 data Location = Span SourcePos SourcePos
               deriving (Eq, Ord, Show)
 
+-- | Pretty prints @S1@ of a @'Span' S1 _@ object with 'sourcePosPretty'
+startPosPretty :: Location -> String
+startPosPretty (Span s _) = sourcePosPretty s
+
+-- | Pretty prints @S2@ of a @'Span' _ S2@ object with 'sourcePosPretty'
+endPosPretty :: Location -> String
+endPosPretty (Span _ s) = sourcePosPretty s
+
 -- | The 'Located' datatype adds a source span to the type @a@
 data Located a = At Location a
-               deriving (Eq, Ord, Show)
+               deriving (Eq, Ord, Show, Functor)
 
 -- | The 'located' function adds a source span to a parser.
 located :: (MonadParsec e s m) => m a -> m (Located a)