+++ ChangeLog.md
@@ -4,8 +4,9 @@ Version 1.1.0.0
 ---------------
 
 * Add startPosPretty and endPosPretty function
-* Add Bifunctor, Bifoldable and Bitraversable instances
+* Add Bifunctor, Bifoldable and Bitraversable instances for SExpr
 * Add Base SExpr, Recursive, Corecursive instances (see package recursion-schemes)
+* Add Functor instance for Located
 
 Version 1.0.0.2
 ---------------